Extract translate2query
