+++ violet/test/query_test.dart
@@ -2,9 +2,9 @@
 // Copyright (C) 2020-2024. violet-team. Licensed under the Apache-2.0 License.
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:violet/component/hitomi/hitomi.dart';
 import 'package:violet/component/hitomi/tag_translate.dart';
 import 'package:violet/component/index.dart';
+import 'package:violet/component/query_translate.dart';
 import 'package:violet/settings/settings.dart';
 
 void main() {
@@ -44,12 +44,12 @@ void main() {
 
     test('Hitomi Query To Sql', () {
       Settings.searchPure = false;
-      final result0 = HitomiManager.translate2query(
-          'female:sole_female (lang:korean or lang:n/a)');
-      final result1 = HitomiManager.translate2query(
-          'female:sole_female -(female:mother female:milf)');
-      final result2 = HitomiManager.translate2query(
-          '(lang:korean or lang:n/a) -female:sole_female');
+      final result0 =
+          translate2query('female:sole_female (lang:korean or lang:n/a)');
+      final result1 =
+          translate2query('female:sole_female -(female:mother female:milf)');
+      final result2 =
+          translate2query('(lang:korean or lang:n/a) -female:sole_female');
 
       expect(result0,
           'SELECT * FROM HitomiColumnModel WHERE Tags LIKE \'%|female:sole female|%\' AND (Language LIKE \'%korean%\' OR Language LIKE \'%n/a%\')  AND ExistOnHitomi=1');
